backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yt-dlp==2024.3.10
python-multipart==0.0.9
slowapi==0.1.9
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
aiofiles==23.2.1
